refactor : 컨트롤러 메소드 id타입 변경, 카프카 예외처리 분기 변수 추가, 예외처리 추가

+++ user-service/src/main/java/com/local_dating/user_service/presentation/controller/CardController.java
@@ -19,19 +19,19 @@ public class CardController {
 
     @PreAuthorize("isAuthenticated() and #id == authentication.getPrincipal()")
     @GetMapping(value = "/v1/users/{id}/cards")
-    public List viewRecomcard(final @PathVariable("id") Integer id, final Authentication authentication) {
+    public List viewRecomcard(final @PathVariable("id") long id, final Authentication authentication) {
         return userCardService.getCard(authentication.getPrincipal().toString());
     }
 
     @PreAuthorize("isAuthenticated() and #id == authentication.getPrincipal()")
     @PostMapping(value = "/v1/users/{id}/cards")
-    public void saveRecomcard(final @PathVariable("id") Integer id, final Authentication authentication) {
+    public void saveRecomcard(final @PathVariable("id") long id, final Authentication authentication) {
         userCardService.setCard(authentication.getPrincipal().toString());
     }
 
     @PreAuthorize("isAuthenticated() and #id == authentication.getPrincipal()")
     @GetMapping(value = "/v1/users/{id}/cards/detail/{targetId}")
-    public List viewRecomcardDetail(final @PathVariable("id") Integer id, final Authentication authentication, @PathVariable String targetId) {
+    public List viewRecomcardDetail(final @PathVariable("id") long id, final Authentication authentication, @PathVariable String targetId) {
         return userCardService.getCardDetail(authentication.getPrincipal().toString(), targetId);
     }
 }
